added reboot command, extra conditionals to decide if a reboot in nececary

+++ monitor_check_startup.sh
@@ -1,15 +1,26 @@
 #!/bin/sh
 
+xrandr --listactivemonitors "$monitorcheck"
+FILE=/etc/X11/xorg.conf.bkp
+
 if [[$monitorcheck == *"Failed to get size of gamma"]]; then
-     sudo mv /etc/X11/xorg.conf.bkp /etc/X11/xorg.conf
-    echo "No monitor detected,Switching to Headless mode!"
+    if test -f "$FILE"; then
+        sudo mv /etc/X11/xorg.conf.bkp /etc/X11/xorg.conf
+        echo "Switching to Headless mode!, rebooting in 5 seconds"
+        sleep 5s
+        sudo reboot
+    else 
+        echo "No monitor detected,xorg.conf detected Headless mode! already enabled"
     sleep 5s
 
 else
-    sudo mv /etc/X11/xorg.conf /etc/X11/xorg.conf.bkp
-    echo "Monitor detected, Switching to Monitor mode!"
-    sleep 5s
+    if test -f "$FILE"; then
+        echo "monitor detected,xorg.conf not detected monitor mode! already enabled"
+        sleep 5s
+    else
+        sudo mv /etc/X11/xorg.conf.bkp /etc/X11/xorg.conf
+        echo "Monitor detected, Switching to Monitor mode!"
+        sleep 5s
+        sudo reboot
 
 fi
-
-xrandr --listactivemonitors "$monitorcheck"
